Add `Traversable#prepare_entrance!`
Connecting the first room in the dungeon before was done with
`Traversable#connect_to` with `nil` passed as an argument.

This limited the entrance room to have `Room#@has_ahead` true value and
for a simple task to be done way later after a bunch of checks.

As a result `Traversable#connect_to` now doesn't accept nil as a
value anymore as well.